fix for HA master switch for SpringTransactionManager integration
SpringTransactionManager and UserTransactionImpl stored a reference to the TransactionManager which was no longer
valid after a master-switch, now the tx-manager is looked up in the GraphDatabaseAPI for all operations.

NOTE: The API of UserTransactionImpl changed, it takes a GraphDatabaseAPI constructor parameter again (was changed in 1.7
to javax.transaction.TransactionManager).

Conflicts:

	kernel/src/main/java/org/neo4j/kernel/impl/transaction/SpringTransactionManager.java
	kernel/src/main/java/org/neo4j/kernel/impl/transaction/UserTransactionImpl.java