--Second attempt, previous attempt a1be8198abc3073280f4538dd87b90adefe6c88c  --broke the build and got reverted by  --014246e540f65e1c0a5aede6245dff5fcd4cab10.
Sessions now live on the server, which means it can hold any type of
objects, not just Serializables. Needed for coming features, such as
holding results from cypher queries.

One drawback of this is that a clients session variables will be lost on
reconnect to the server, however initial session variables (i.e. given
when starting the client) will be re-set in the new session after a
reconnect.