Resolve compile error with Oracle JDK 7.
This resolves the following complie error:

/home/joe/code/neo4j/community/graphviz/src/main/java/org/neo4j/visualization/graphviz/Script.java:[116,14]
error: storeDir has private access in Script

As far as I can tell the reason is that S is a generic type, which
extends from Script, so as a subclass of Script, S only has no access to
private members of Script even if S is of type Script rather than a
subclass. The bug report against JDK7 was rejected, saying this is
correct behaviour, meaning the JDK6 was wrong to accept this.

For an more thorough explanation why this is needed see:
http://stackoverflow.com/questions/10782876/changes-in-access-of-variables-for-generic-classes-in-java-7/10798968#10798968

The bug report against JDK7:
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=7022052