PersistenceWindowPool (and friends) bug fixes and cleanup:
o Proper closing of windows, not just closing for the sake of preventing
marking as in use in all places.
o Proper handover of PersistenceRows to other threads if marked deciding
whether to close or not.
o Simplified state in PersistenceRow and LockableWindow.
o Reimplemented releasing a PersistenceRow to force unlock being the last
operation
o PersistenceRows are marked as in use in the constructor to avoid an
issue where a row could be shared too early
o activeRowWindows map has got Long as key instead of Integer to prevent
wrap-around of big integers.
o Fixed a potential problem in LockableWindow#unLock where a locking
thread might have registered itself more than once in a row preventing
another thread waiting in line to lock to become notified (via interrupt)

Initially we set out to fix BufferOverflow/UnderflowExceptions arising from
closed or contemporary closing PersistenceRow that leaked out by mistake.
In the meantime we cleaned up the code and fixed other problems that the
changes exposed.

co-authors: Alistair Jones, Tobias Lindaaker